refactor(image card section): modularize codebase
fixes #42
